requirements.txt
# Core
pillow>=10.0.0
pillow-heif>=0.15.0
pyexiftool>=0.5.6
pyyaml>=6.0

# Web UI
flask>=3.0.0

# Windows timestamps (optional but recommended on Windows)
pywin32>=306; sys_platform == "win32"
